+++ src/tools/mod.rs
@@ -1,8 +1,9 @@
 pub mod current_datetime;
 pub mod filesystem;
 pub mod parallel_tasks;
+pub mod ralph_wiggum_loop;
 pub mod search;
-// pub mod shell;
+pub mod shell;
 pub mod soul;
 pub mod state_manager;
 pub mod system_status;
@@ -10,3 +11,4 @@ pub mod todo;
 pub mod weather;
 pub mod web_fetch;
 pub mod wiki;
+